[OD19DEVAT1-78] - changes in path

+++ src/main/java/com/netcracker/services/PredictionService.java
@@ -1,12 +1,16 @@
 package com.netcracker.services;
 
+import com.netcracker.models.MonthReport;
+
 import java.math.BigInteger;
+import java.util.List;
 
 public interface PredictionService {
 
-    public boolean predictCreditPossibility(BigInteger id, int duration, long amount);
+    public boolean predictCreditPossibility(BigInteger id, int duration, double amount);
 
     public double predictMonthIncome(BigInteger id, int duration);
 
     public double predictMonthExpense(BigInteger id, int duration);
+
 }
